make the f128 exponent explicitly int64
The exponent was a platform-sized int, which is 32 bits on 386/arm. No reachable input wraps it -- consensus parameters keep every exponent in the low thousands, and even the unfiltered fuzz domain stays near 2*10^5 -- but that made the cross-GOARCH bit-identity guarantee depend on magnitudes staying small rather than on the type. Also restructure add's alignment check to compare exponents in int64 BEFORE converting to a uint shift count: uint is likewise 32-bit on those platforms, and the naive type swap would leave a truncation that could alias a huge exponent difference to a small shift.

+++ f128_test.go
@@ -298,8 +298,8 @@ func FuzzF128Ops(f *testing.F) {
 	f.Add(uint64(1)<<63, uint64(0), 0, uint64(3)<<62, uint64(0), 0, uint64(7))
 	f.Add(uint64(0), uint64(0), 0, uint64(1)<<63, uint64(1), -5, uint64(1))
 	f.Fuzz(func(t *testing.T, ahi, alo uint64, aexp int, bhi, blo uint64, bexp int, u uint64) {
-		a := norm128(ahi, alo, aexp%4000-2000)
-		b := norm128(bhi, blo, bexp%4000-2000)
+		a := norm128(ahi, alo, int64(aexp%4000-2000))
+		b := norm128(bhi, blo, int64(bexp%4000-2000))
 		ab, bb := f128ToBig(a), f128ToBig(b)
 		check := func(name string, got f128, want *big.Float) {
 			gotBig := f128ToBig(got)
@@ -361,7 +361,7 @@ func f128ToBig(x f128) *big.Float {
 	hi := new(big.Float).SetPrec(300).SetUint64(x.hi)
 	hi.SetMantExp(hi, 64)
 	m := new(big.Float).SetPrec(300).Add(hi, new(big.Float).SetPrec(300).SetUint64(x.lo))
-	return m.SetMantExp(m, x.exp)
+	return m.SetMantExp(m, int(x.exp))
 }
 
 // TestDivVsBig checks f128.div against a 128-bit round-nearest-even big.Float
@@ -372,7 +372,7 @@ func f128ToBig(x f128) *big.Float {
 func TestDivVsBig(t *testing.T) {
 	rng := rand.New(rand.NewSource(2))
 	randF128 := func() f128 {
-		return f128{rng.Uint64() | 1<<63, rng.Uint64(), rng.Intn(4000) - 2000} // normalized
+		return f128{rng.Uint64() | 1<<63, rng.Uint64(), int64(rng.Intn(4000) - 2000)} // normalized
 	}
 	for i := 0; i < 5_000_000; i++ {
 		a, b := randF128(), randF128()
